[#118] cleaner way to create system tests
