+++ src/main/java/frc/robot/command/tests/DrivetrainSystemTest.java
@@ -6,26 +6,30 @@ package frc.robot.command.tests;
 
 import com.ctre.phoenix6.mechanisms.swerve.SwerveRequest;
 
-import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
+import frc.robot.command.tests.testCommands.DriveTest;
 import frc.robot.subsystems.Swerve;
 import frc.thunder.command.TimedCommand;
+import frc.thunder.testing.SystemTestCommandGroup;
 
-public class DrivetrainSystemTest extends SequentialCommandGroup {
-
+public class DrivetrainSystemTest extends SystemTestCommandGroup {
+  
   public DrivetrainSystemTest(Swerve drivetrain, SwerveRequest brake, double speed) {
-    addCommands(
-      new WaitCommand(0.5),
-      new TimedCommand(new DriveTest(drivetrain, () -> speed, () -> 0d), 1), // Forward
-      new WaitCommand(1),
-      new TimedCommand(new DriveTest(drivetrain, () -> -speed, () -> 0d), 1), // Backward
-      new WaitCommand(1),
-      new TimedCommand(new DriveTest(drivetrain, () -> 0d, () -> speed), 1), // Left
-      new WaitCommand(1),
-      new TimedCommand(new DriveTest(drivetrain, () -> 0d, () -> -speed), 1), // Right
-      new WaitCommand(0.5),
-      drivetrain.applyRequest(() -> brake) // Brake
+    super(
+      new SequentialCommandGroup(
+        new WaitCommand(0.5),
+        new TimedCommand(new DriveTest(drivetrain, () -> speed, () -> 0d), 1), // Forward
+        new WaitCommand(1),
+        new TimedCommand(new DriveTest(drivetrain, () -> -speed, () -> 0d), 1), // Backward
+        new WaitCommand(1),
+        new TimedCommand(new DriveTest(drivetrain, () -> 0d, () -> speed), 1), // Left
+        new WaitCommand(1),
+        new TimedCommand(new DriveTest(drivetrain, () -> 0d, () -> -speed), 1), // Right
+        new WaitCommand(0.5),
+        drivetrain.applyRequest(() -> brake) // Brake
+      )
     );
   }
+
 }
